requirements.txt
bctpy==0.4.1
connectome-utils==0.0.0
cycler==0.9.0
decorator==4.0.4
hiveplotter==0.1
ipython==4.0.0
ipython-genutils==0.1.0
matplotlib==1.5.0
networkx==1.10
numpy==1.10.1
pandas==0.17.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.0.0
ptyprocess==0.5
pyparsing==2.0.5
python-dateutil==2.4.2
pytz==2015.7
PyX==0.14.1
scipy==0.16.1
seaborn==0.6.0
Shapely==1.5.13
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
wheel==0.24.0
